refactor(codificação): separa storage, sessão e ciclo de vida do rascunho (#608)
Decompõe useCodingDrafts em três camadas — lib pura, acesso ao localStorage +
máquina de sessão sem React, e o hook de ciclo de vida — e extrai a fiação da
tela para useCodingDraftWiring. Motivado pelo gate de complexidade do fallow.
